Primary node clarification for VK_EXT_physical_device_drm
The extension hinted that the primary node represents display
capabilities as its intended purpose, but the current text let it
ambiguous enough that there has been confusion and proposed
implementations which expose different devices for the primary node.

Add clarification notes to further define the scope of primary node
in the context of this extension.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>